scraper/requirements.txt
python-dotenv
supabase
requests
feedparser
numpy<2.0
transformers
torch
nba-api
safetensors
pandas